# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  service.OTPSendRequest:
    properties:
      device_fingerprint:
        type: string
      device_id:
        type: string
      ip_address:
        type: string
      phone_number:
        maxLength: 15
        minLength: 10
        type: string
      provider:
        type: string
      purpose:
        enum:
        - login
        - verification
        - password_reset
        - admin_login
        - forgot_mpin
        type: string
      user_agent:
        type: string
    required:
    - phone_number
    - purpose
    type: object
  service.OTPVerifyRequest:
    properties:
      device_fingerprint:
        type: string
      device_id:
        type: string
      ip_address:
        type: string
      otp:
        type: string
      phone_number:
        maxLength: 15
        minLength: 10
        type: string
      purpose:
        enum:
        - login
        - verification
        - password_reset
        - admin_login
        - forgot_mpin
        type: string
      user_agent:
        type: string
    required:
    - otp
    - phone_number
    - purpose
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Auth Service for B2B Suite – admin, employee, attendance, etc.
  title: B2B Suite API
  version: "1.0"
paths:
...
  /api/v1/auth/otp/send:
    post:
      consumes:
      - application/json
      description: Sends OTP for login, verification, password reset, or forgot MPIN.
      parameters:
      - description: OTP send request
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OTP sent
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "429":
          description: Rate limit or quota exceeded
          schema:
            additionalProperties: true
            type: object
      summary: Send OTP
      tags:
      - auth-user
...
  /api/v1/otp/cleanup:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: Cleanup result
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Cleanup failed
          schema:
            additionalProperties: true
            type: object
      summary: Cleanup expired OTPs
      tags:
      - otp
...
  /api/v1/otp/send:
    post:
      consumes:
      - application/json
      description: Sends an OTP to the provided phone number.
      parameters:
      - description: OTP send request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.OTPSendRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP sent
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation failed
          schema:
            additionalProperties: true
            type: object
        "429":
          description: Rate limit exceeded
          schema:
            additionalProperties: true
            type: object
      summary: Send OTP
      tags:
      - otp
  /api/v1/otp/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: Statistics
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to get stats
          schema:
            additionalProperties: true
            type: object
      summary: Get OTP stats
      tags:
      - otp
  /api/v1/otp/verify:
    post:
      consumes:
      - application/json
      description: Verifies the OTP and can issue authentication tokens.
      parameters:
      - description: OTP verification request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.OTPVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Verification result with tokens
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation failed
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid OTP
          schema:
            additionalProperties: true
            type: object
        "429":
          description: Too many attempts
          schema:
            additionalProperties: true
            type: object
      summary: Verify OTP
      tags:
      - otp
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"